test: per-bar canvas assertions for redundant encoding example
Replace the whole-canvas ink assertion in pattern-scale-test.js with
per-bar ink sampling using the absoluteBounds/countInk machinery from
pattern-legend-test.js, now extracted to a shared test/util.js (matching
the test/util.js precedent used by vega-format, vega-time, vega-cli).
Also trims the pattern-scale-redundant.vg.json description down to the
accessibility claim.
